Added the ability to call jenkins_create_project script with arguments
