+++ CES Service Manager/CES Service Manager.csproj
@@ -76,10 +76,10 @@
     <Compile Include="Form_EnvironmentJenkins.Designer.cs">
       <DependentUpon>Form_EnvironmentJenkins.cs</DependentUpon>
     </Compile>
-    <Compile Include="Form_Jenkins_CreateProject.cs">
+    <Compile Include="Forms\Form_Jenkins_CreateProject.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form_Jenkins_CreateProject.Designer.cs">
+    <Compile Include="Forms\Form_Jenkins_CreateProject.Designer.cs">
       <DependentUpon>Form_Jenkins_CreateProject.cs</DependentUpon>
     </Compile>
     <Compile Include="Form_Main.cs">
@@ -114,7 +114,7 @@
     <EmbeddedResource Include="Form_EnvironmentJenkins.resx">
       <DependentUpon>Form_EnvironmentJenkins.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Form_Jenkins_CreateProject.resx">
+    <EmbeddedResource Include="Forms\Form_Jenkins_CreateProject.resx">
       <DependentUpon>Form_Jenkins_CreateProject.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Form_Main.resx">
