+++ CES Service Manager/Forms/Form_Jenkins_CreateProject.resx
